🗑️ Chore : 폴더 구조 변경 및 사진추가 컴포넌트 분리
