+++ src/pages/PostCreate/PostContent/PostContent.css
