+++ src/pages/PostCreate/PostContent/PostContent.tsx
